+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Provide an option to compact the title screen (reposition elements / remove redundant buttons)
     * Skip friendlist promotion
     * Integrated server: Disable `enforce-secure-profile` by default
+* Removed legacy configuration directory migration
 * Minor performance improvements and tweaks
 
 # 1.7.0